allow fetch tests only for meteor 1.7+
